+++ role_play/settings/base.py
@@ -38,6 +38,15 @@ ROOT_URLCONF = "role_play.urls"
 LOGIN_REDIRECT_URL = "/game"
 LOGOUT_REDIRECT_URL = "/game"
 
+# Password hashers
+PASSWORD_HASHERS = [
+    "django.contrib.auth.hashers.ScryptPasswordHasher",
+    "django.contrib.auth.hashers.PBKDF2PasswordHasher",
+    "django.contrib.auth.hashers.PBKDF2SHA1PasswordHasher",
+    "django.contrib.auth.hashers.Argon2PasswordHasher",
+    "django.contrib.auth.hashers.BCryptSHA256PasswordHasher",
+]
+
 # Password validation
 AUTH_PASSWORD_VALIDATORS = [
     {